src/lib/tests/Generator.class.test.ts
import Generator from "../Generator.class.js";
import { EventEmitter } from 'events';
import Stage from "../Stage.class.js";
import parseYamlFile from "../../utils/parseYamlFile.js";
import Pipeline from "../Pipeline.class.js";
import * as chai from 'chai'
import chaiAsPromised from 'chai-as-promised'
import { NamedNode } from "n3";
import type { LDWorkbenchConfiguration } from "../LDWorkbenchConfiguration.js";
import * as fs from 'fs';
chai.use(chaiAsPromised)
const expect = chai.expect


// function compareFiles(file1Path: string, file2Path: string): boolean {
//   const file1Content = fs.readFileSync(file1Path, 'utf-8');
//   const file2Content = fs.readFileSync(file2Path, 'utf-8');

//   const file1Lines = file1Content.split('\n').sort();
//   const file2Lines = file2Content.split('\n').sort();

//   const areLinesEqual = JSON.stringify(file1Lines) === JSON.stringify(file2Lines);

//   if (!areLinesEqual) {
//     const diffLines = file1Lines.filter(line => !file2Lines.includes(line));
//     console.log('🪵  | file: Generator.class.test.ts:28 | compareFiles | diffLines:', diffLines)
//     // changed to console.error so test won't throw in CI
//     console.error(`Files are different. Lines in ${file1Path} that are not in ${file2Path}:\n${diffLines.join('\n')}`);
//   }

//   return true;
// }


describe.only('Generator Class', () => {
    describe.skip('constructor', () => {
        it('should set query, engine, endpoint, and source properties correctly', () => {
            const configuration = parseYamlFile('./static/example/config.yml')
            const pipeline = new Pipeline(configuration)
            const stageConfig = configuration.stages[0]
            const stage = new Stage(pipeline, stageConfig)
            const generator = new Generator(stage)
            expect(generator).to.be.an.instanceOf(Generator);
            expect(generator).to.be.an.instanceOf(EventEmitter);
            expect(generator).to.have.property('query');
            expect(generator).to.have.property('engine');
            expect(generator).to.have.property('endpoint');
            expect(generator).to.have.property('source');
        });
    });
    // BUG when both the generator and iterator tests are running, it seems the iterator will never terminate
    describe('run', () => {
        // @mightymax uncomment this to see File class returning a resolve before it finished
        // afterEach(async ()=>{
        //     async function removeFile(filePath: string): Promise<void> {
        //         return new Promise((resolve, reject) => {
        //           // Use fs.unlink to remove the file
        //           fs.unlink(filePath, (err) => {
        //             if (err != null) {
        //               reject(err);
        //             } else {
        //               resolve();
        //             }
        //           });
        //         });
        //       }
            
        //     const filePath = 'src/lib/tests/data/example-pipelineBatch.nt';
        //     if(fs.existsSync(filePath)){
        //         await removeFile(filePath)
        //     }
        // })
        it.only('Should work in single batch pipeline', async function (){
            this.timeout(3000)
            const filePath = 'src/lib/tests/data/example-pipelineBatch.nt';


            const batchConfiguration: LDWorkbenchConfiguration = {
                name: 'Example Pipeline Batch',
                description: 'This is an example pipeline. It uses files that are available in this repository  and SPARQL endpoints that should work.\n',
                destination: "file://"+filePath,
                stages: [
                  {
                    name: 'Stage 1',
                    iterator: {
                      query: 'file://static/example/iterator-stage-1.rq',
                      endpoint: 'file://static/test/iris.nt'
                    //   endpoint: 'file://static/test/iris-small.nt'
                    },
                    generator: {
                      query: 'file://static/example/generator-stage-1.rq',
                      // adjust batchsize for test here
                      batchSize: 3
                    }
                  }
                ]
              }

              const pipelineBatch = new Pipeline(batchConfiguration)
              async function runPipelineWithPromise(): Promise<boolean> {
                let batchPipelineEnd = false
                return new Promise((resolve, reject) => {
                            pipelineBatch.run().then(_ => {
                                // waiting for the "end" event to be emitted
                                pipelineBatch.on('end', () => {
                                    // @mightymax even the "end" event is emitted in Pipeline (which should happen when writeResult has finished), the file seems to not have finished writing
                                    // on the first run it cannot find the file, since it has not been written
                                    // on the second run it will show the file it created from the run before

                                    // if you uncomment out the afterEach for these tests, it will delete the file and you will notice that it will fail each time

                                    // when iris.nt (the endpoint) is small (8 instances) it seems never to be able to write the file with batch processing
                                    // => this can be  tested by using iris-small.nt in the configuration above
                                    
                                    // notice that the byte size is significantly larger for batch processing than for single processing
                                    // the byte size of the buffer is also inconsistent for batch, but becomes consistent when batch is not used
                                    // see: src/lib/File.class.ts
                                    try {                                        
                                        const fileContent = fs.readFileSync(filePath, {encoding:'utf8'});
                                        const fileLines = fileContent.split('\n').sort();
                                        console.log('🪵  | file: Generator.class.test.ts:106 | pipelineBatch.addListener | fileLines:', fileLines)
                                        batchPipelineEnd = true 
                                        if (batchPipelineEnd){
                                            resolve(true)
                                        }
                                    } catch (error) {
                                        console.error(error)
                                    }
                                });
                            }).catch(e => {reject(e)})
                });
            }
            await runPipelineWithPromise()


        })
        // it.skip('Should work with batch processing: normal pipeline compared to batch pipeline', async function (){
        //     this.timeout(200000)
        //     const file1Path = 'pipelines/data/example-pipeline.nt';
        //     const file2Path = 'pipelines/data/example-pipelineBatch.nt';

        //     const configuration: LDWorkbenchConfiguration = {
        //         name: 'Example Pipeline',
        //         description: 'This is an example pipeline. It uses files that are available in this repository  and SPARQL endpoints that should work.\n',
        //         destination: "file://"+file1Path,
        //         stages: [
        //           {
        //             name: 'Stage 1',
        //             iterator: {
        //               query: 'file://static/example/iterator-stage-1.rq',
        //               endpoint: 'file://static/test/iris.nt'
        //             },
        //             generator: {
        //               query: 'file://static/example/generator-stage-1.rq',
        //             }
        //           }
        //         ]
        //       }
        //     const batchConfiguration: LDWorkbenchConfiguration = {
        //         name: 'Example Pipeline Batch',
        //         description: 'This is an example pipeline. It uses files that are available in this repository  and SPARQL endpoints that should work.\n',
        //         destination: "file://"+file2Path,
        //         stages: [
        //           {
        //             name: 'Stage 1',
        //             iterator: {
        //               query: 'file://static/example/iterator-stage-1.rq',
        //               endpoint: 'file://static/test/iris.nt'
        //             },
        //             generator: {
        //               query: 'file://static/example/generator-stage-1.rq',
        //               // adjust batchsize for test here
        //               batchSize: 4
        //             }
        //           }
        //         ]
        //       }

        //       const pipeline = new Pipeline(configuration)
        //       const pipelineBatch = new Pipeline(batchConfiguration)
        //       async function runPipelineWithPromise(): Promise<boolean> {
        //         let pipelineEnd = false
        //         let batchPipelineEnd = false
        //         return new Promise((resolve, reject) => {
        //             pipeline.addListener('error', (error) => {
        //                 reject(error);
        //             });
        //             // running the normal pipeline
        //             pipeline.run().then(_ => {
        //                 // waiting for the "end" event to be emitted
        //                 pipeline.on('end', () => {
        //                     pipelineEnd = true
        //                     // running the batch pipeline
        //                     pipelineBatch.run().then(_ => {
        //                         // waiting for the "end" event to be emitted
        //                         pipelineBatch.on('end', () => {
        //                             batchPipelineEnd = true 

        //                             if (pipelineEnd && batchPipelineEnd){
        //                                 resolve(true)
        //                             }
        //                         });
        //                     }).catch(_ => {})
        //                 });
        //             }).catch(_ => {})
        //         });
        //     }
        //     await runPipelineWithPromise()
        //     // BUG should return true but doesnt due to the File class's write() which returns a smaller buffer size for the batch pipeline
        //     // expect(compareFiles(file1Path, file2Path)).to.equal(true)
        //     compareFiles(file1Path, file2Path)
        // }
        // )
        it.skip('should emit "data" and "end" events with the correct number of statements', async () => {
            const configuration = parseYamlFile('./static/example/config.yml')
            const pipeline = new Pipeline(configuration)
            const stageConfig = configuration.stages[0]
            const stage = new Stage(pipeline, stageConfig)
            const generator = new Generator(stage);
            const emittedEvents: any[] = [];

            const testNamedNode = new NamedNode('https://triplydb.com/triply/iris/id/floweringPlant/00106');

            async function runGeneratorWithPromise(): Promise<boolean> {
                return new Promise((resolve, reject) => {
                    generator.addListener('data', (quad) => {
                        emittedEvents.push({ event: 'data', quad });
                    });
                    generator.addListener('end', (numResults) => {
                        emittedEvents.push({ event: 'end', numResults });
                        resolve(true);
                    });
                    generator.addListener('error', (error) => {
                        reject(error);
                    });
                    generator.run(testNamedNode);
                });
            }

            await runGeneratorWithPromise()
            expect(emittedEvents).to.have.lengthOf(4);
            expect(emittedEvents[0].event).to.equal('data');
            expect(emittedEvents[0].quad._subject.id).to.equal('https://triplydb.com/triply/iris/id/floweringPlant/00106')
            expect(emittedEvents[0].quad._predicate.id).to.equal('http://www.w3.org/1999/02/22-rdf-syntax-ns#type')
            expect(emittedEvents[0].quad._object.id).to.equal('https://schema.org/Thing')
            expect(emittedEvents[emittedEvents.length - 1].event).to.equal('end');
            expect(emittedEvents[emittedEvents.length - 1].numResults).to.equal(3);
        });
    });
});
